+++ src/main/java/restaurant/manager/RestaurantWebServiceWithDatabase/Services/RestaurantService.java
@@ -4,7 +4,10 @@ import lombok.AllArgsConstructor;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 import restaurant.manager.RestaurantWebServiceWithDatabase.Entities.Restaurant;
+import restaurant.manager.RestaurantWebServiceWithDatabase.Entities.User;
+import restaurant.manager.RestaurantWebServiceWithDatabase.Exceptions.NotFoundException;
 import restaurant.manager.RestaurantWebServiceWithDatabase.Repositories.RestaurantRepository;
+import restaurant.manager.RestaurantWebServiceWithDatabase.Repositories.UserRepository;
 
 import java.util.List;
 
@@ -13,6 +16,7 @@ import java.util.List;
 public class RestaurantService {
 
     private RestaurantRepository restaurantRepository;
+    private UserRepository userRepository;
 
     public List<Restaurant> fetchAllRestaurants() {
         return restaurantRepository.findAllRestaurantsJoinFetchFoodsOwner();
@@ -22,7 +26,10 @@ public class RestaurantService {
         return restaurantRepository.findByRestaurantIdJoinFetchFoodsOwner(id);
     }
 
-    public void addOneRestaurant(@NonNull Restaurant restaurant) {
+    public void addOneRestaurant(@NonNull Restaurant restaurant,
+                                 @NonNull Integer ownerId) {
+        User owner = userRepository.findByUserId(ownerId);
+        restaurant.setOwner(owner);
         restaurantRepository.save(restaurant);
     }
 
@@ -31,6 +38,9 @@ public class RestaurantService {
     }
 
     public void removeOneRestaurant(@NonNull Integer id) {
+        if (!restaurantRepository.existsById(id)) {
+            throw new NotFoundException("No Restaurant was found to be deleted");
+        }
         restaurantRepository.deleteById(id);
     }
 }
